requirements.txt
pandas
vnlunar
vncalendar
python-telegram-bot
fastapi
uvicorn
python-multipart
requests
tabulate
gspread
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
